tests: move method to mock iterators from TestCase to trait
