+++ test/TestCase.php
@@ -18,17 +18,13 @@
  * Copyright (c) 2018-2021 (original work) Open Assessment Technologies SA;
  */
 
-declare(strict_types=1);
-
 namespace oat\generis\test;
 
-use stdClass;
+use oat\oatbox\service\ServiceManager;
 use Prophecy\Argument;
 use Psr\Container\ContainerInterface;
-use oat\oatbox\service\ServiceManager;
-use PHPUnit\Framework\MockObject\MockObject;
-use PHPUnit\Framework\TestCase as UnitTestCase;
 use Zend\ServiceManager\ServiceLocatorInterface;
+use PHPUnit\Framework\TestCase as UnitTestCase;
 
 abstract class TestCase extends UnitTestCase
 {
@@ -59,82 +55,5 @@ abstract class TestCase extends UnitTestCase
         $containerProphecy->has(Argument::any())->willReturn(false);
 
         return $serviceLocatorProphecy->reveal();
-    }
-
-    public function createIteratorMock(string $originalClassName, array $items = []): MockObject
-    {
-        $iteratorData = new stdClass();
-        $iteratorData->array = $items;
-        $iteratorData->position = 0;
-
-        $iteratorMock = $this->createMock($originalClassName);
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('rewind')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        $iteratorData->position = 0;
-                    }
-                )
-            );
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('current')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        return $iteratorData->array[$iteratorData->position];
-                    }
-                )
-            );
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('key')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        return $iteratorData->position;
-                    }
-                )
-            );
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('next')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        $iteratorData->position++;
-                    }
-                )
-            );
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('valid')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        return isset($iteratorData->array[$iteratorData->position]);
-                    }
-                )
-            );
-
-        $iteratorMock
-            ->expects($this->any())
-            ->method('count')
-            ->will(
-                $this->returnCallback(
-                    static function() use ($iteratorData) {
-                        return sizeof($iteratorData->array);
-                    }
-                )
-            );
-
-        return $iteratorMock;
     }
 }
